Use default resolution of 1280x720 when video size is unknown. Fixes #1571.
This is more or less a workaround, since 1280x720 is hard-coded.
But I think there is no need to do anything more complicated for
now (like adding a way to specify the default resolution, through
a primitive) because the root cause of this is obvioulsy a libvlc
bug which will (hopefully) be fixed.