Do not return texture until video has actully started playing. Refs #1915.
This commit fixes a side-effect of commit b734125: when playing a
video, the first image is not a full frame, but a solid color. It is
especially visible when streaming a YouTube video. This occurs when we play
the video with a 1x1 reesolution to probe for the real resolution. To avoid
the problem we make sure we don't copy the frame to a GL texture when we are
probing.