Use Pixel Buffer Objects (PBOs) when available. Closes #2351.
This commit adds support for PBOs to improve performance.
If the GL version is 2.1 or above, the transfer of the pixel data
to the GPU can be made essentially asynchronous. Instead of converting
the picture from VLC to a suitable GL format into a regular memory buffer,
then calling glTexImage2D to transfer it into the texture memory, we use
PBOs to map GL memory into the address space of the process. The image
decoded by VLC is converted into the proper format directly into the
mapped memory, then the GPU starts the transfer asynchronously wrt.
the CPU. When a frame is being transfered, the previous one is displayed
(this is called "ping-pong" mode, because two PBOs are used).

When available, PBOs are used by default. PBO usage can be disabled by
setting the environment variable TAO_VLC_NO_PBO to 1.
On MacOSX, VLC chroma can be forced to RV32 with TAO_VLC_RV32=1
(default is to use UYVY).

Here are some test results (MacBook Pro 2010, 2.8 GHz Core 2 Duo, NVIDIA
9600M GT). The test document shows four videos simultaneously. The video
source is a 1280x720 .mp4. The same file is played four times by
prepending "./" to the file path. Tao window is 1620x960, each video
occupies 540x310 pixels. The video textures have the native video resolution
(1280x720). VLC is initialized with:
  vlc_init "--ffmpeg-threads", "1", "--noaudio"
The table shows frames per seconds and CPU usage (fps, CPU%).

                     RV32              UYVY
Without PBOs     (7, 180%)        (35, 160%)
With PBOs       (10, 170%)        (56, 130%)

On this test case, the frame rate is improved by 60% and CPU usage is
reduced by almost 20% compared to the non-PBO case.