Compensate for audio/video delay introduced by PBO ping-pong, when FPS are known. Refs #2351.
Since we are always one frame late when using PBOs to transfer images to
the GL texture, we should also delay the audio by the same amount. This is
easily done when the frame rate is known. This commit does exactly that.

In case the FPS value is unknown at the time we get the video
format callback, we could still measure the actual frame rate in the
displayFrame callback. I'm not addressing this case for now.