Fix errors on Windows and Linux. Refs #2351.
Fixes a compilation error on Windows and a crash on Linux.
On Windows and Linux we use GLEW. Unfortunately it was initialized too
soon, returning an error "Missing GL version" (ignored) and so some
GL function pointers were NULL, resulting in a crash as we were trying to
create PBOs.