VlcVideoSurface: keep track of allocated video buffers for proper cleanup. Fixes #2771.
When a video is discarded with movie_drop or movie_only, the LibVLC
player is stopped and deleted. If this occurs during playback (as
opposed to when the movie is done playing), we typically still own one
or more video buffers, allocated by VlcVideoSurface::lockFrame() and
not yet freed (because the frames were never displayed).
This commit uses a QSet<void *> to keep track of allocated memory and
cleans up in the destructor.