VLCAudioVideo: MacOSX: look for VLC under module path first. Fixes #1944, refs #2738.
This commit changes the order in which directories are scanned when looking
for VLC libraries on MacOSX. The module path now takes precedence over the
application path (/Applications/VLC.app). The previous policy introduced by

This commit avoids bug #1944. We shall probably revert to the previous
search order when VLC finally ships with FontConfig >= 2.10.91 on MacOSX.