Do not pass texture size to Tao until a picture is actually available. Fixes #3075.
Test case: without this fix, the following code would show a white, untextured
rectangle for a fraction of a second before showing the actual video.
This is because the 'video size available' and the 'new picture available'
notifications are separate callbacks, so there is an interval of time during
which the size is known but no texture is available, yet.

clear_color 0, 0, 0, 1
color "white"
movie_texture "movie.mp4"
rectangle texture_width, texture_height